添加分页参数解析 组维护 添加事务 添加注解条件查询

+++ demodb/src/main/java/com/yjt/springcloud/demodb/service/impl/RoleServiceImpl.java
@@ -1,17 +1,24 @@
 package com.yjt.springcloud.demodb.service.impl;
 
 import com.baomidou.mybatisplus.core.toolkit.IdWorker;
+import com.google.common.collect.Lists;
+import com.yjt.springcloud.demodb.entity.Org;
 import com.yjt.springcloud.demodb.entity.Role;
 import com.yjt.springcloud.demodb.entity.RoleDataPermission;
-import com.yjt.springcloud.demodb.repository.OrgRepository;
-import com.yjt.springcloud.demodb.repository.RoleDataPermissionRepository;
-import com.yjt.springcloud.demodb.repository.RoleRepository;
+import com.yjt.springcloud.demodb.entity.UserRole;
+import com.yjt.springcloud.demodb.entity.dto.OrgTreeVo;
+import com.yjt.springcloud.demodb.entity.dto.RoleTreeVo;
+import com.yjt.springcloud.demodb.enums.RoleDataScopeEnum;
+import com.yjt.springcloud.demodb.repository.*;
 import com.yjt.springcloud.demodb.service.RoleService;
+import com.yjt.springcloud.demodb.service.mapper.RoleMapper;
 import org.apache.commons.collections4.MapUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 
 /**
  * 角色
@@ -29,6 +36,12 @@ public class RoleServiceImpl implements RoleService {
     private RoleDataPermissionRepository roleDataPermissionRepository;
     @Autowired
     private OrgRepository orgRepository;
+    @Autowired
+    private UserRoleRepository userRoleRepository;
+    @Autowired
+    private UserRepository userRepository;
+    @Autowired
+    private RoleMapper roleMapper;
 
     @Override
     public Role add(Role role) {
@@ -48,7 +61,7 @@ public class RoleServiceImpl implements RoleService {
     @Override
     public Role update(Role role) {
         Role r =  roleRepository.update(role);
-        if(!r.getRoleDataPermissions().isEmpty()){
+        if(!r.getRoleDataPermissions().isEmpty() && RoleDataScopeEnum.CUSTOMER.getType().equals(role.getDataScope())){
             roleDataPermissionRepository.deleteByRole(role.getId());
             role.getRoleDataPermissions().stream().forEach(roleDataPermission -> {
                 roleDataPermission.setId(IdWorker.getId());
@@ -56,6 +69,8 @@ public class RoleServiceImpl implements RoleService {
                 roleDataPermission.setRole(role);
                 roleDataPermissionRepository.save(roleDataPermission);
             });
+        }else{
+            roleDataPermissionRepository.deleteByRole(role.getId());
         }
         return r;
     }
@@ -69,12 +84,83 @@ public class RoleServiceImpl implements RoleService {
     public Role updateStateById(Map param) {
         Long roleId = MapUtils.getLong(param,"id");
         Role role = roleRepository.findById(roleId).get();
-        role.setState(MapUtils.getString(param,"state"));
-        return roleRepository.update(role);
+        List<Role> roles = Lists.newArrayList();
+        roleChildrenList(role,roles);
+        roles.stream().forEach(r -> {
+            r.setState(MapUtils.getString(param,"state"));
+            roleRepository.update(r);
+        });
+
+        return role;
     }
 
     @Override
     public void deleteById(Long roleId) {
-        roleRepository.deleteById(roleId);
+        Role role = roleRepository.findById(roleId).get();
+        List<Role> roles = Lists.newArrayList();
+        roleChildrenList(role,roles);
+        roles.forEach(e->{
+            roleRepository.deleteById(e.getId());
+            List<UserRole> userRoleList = userRoleRepository.findByRole(e);
+            userRoleRepository.deleteAll(userRoleList);
+        });
+
+    }
+
+    @Override
+    public void assginUser(List<Long> userIds, Long roleId) {
+        List<UserRole> userRoleList = Lists.newArrayList();
+        Role role = roleRepository.findById(roleId).get();
+        for (Long userId: userIds) {
+            UserRole userRole = new UserRole();
+            userRole.setId(IdWorker.getId());
+            userRole.setRole(role);
+            userRole.setUser(userRepository.getOne(userId));
+            userRoleList.add(userRole);
+        }
+        if(!userRoleList.isEmpty()){
+            userRoleRepository.saveAll(userRoleList);
+        }
+
+    }
+
+    @Override
+    public RoleTreeVo tree(Map param) {
+        if (MapUtils.getLong(param, "id") != null) {
+            Optional<Role> current = roleRepository.findById(MapUtils.getLong(param, "id"));
+            if(current.isPresent()){
+                Role role=  bulidLeaf(current.get(), Lists.newArrayList());
+                return roleMapper.toDto(role);
+            }
+        } else {
+            Optional<Role> current = roleRepository.findByParentIdIsNull();
+            if(current.isPresent()){
+                Role role=  bulidLeaf(current.get(), Lists.newArrayList());
+                return roleMapper.toDto(role);
+            }
+        }
+        return null;
+    }
+    protected Role bulidLeaf(Role role, List<Role> children) {
+        List<Role> childrens = roleRepository.findByParentId(role.getId());
+        if (!childrens.isEmpty()) {
+            role.setChildren(childrens);
+            childrens.stream().forEach(e -> {
+                bulidLeaf(e, roleRepository.findByParentId(e.getId()));
+            });
+        }
+        return role;
+    }
+
+    protected List<Role> roleChildrenList(Role role,List<Role> children){
+        children.add(role);
+        List<Role> childrens = roleRepository.findByParentId(role.getId());
+        if(!children.isEmpty()){
+            children.addAll(childrens);
+            children.stream().forEach(e->{
+                roleChildrenList(e,roleRepository.findByParentId(e.getId()));
+            });
+        }
+        return children;
     }
 }
